refactor: split studio subsystems
