+++ packages/studio/src/assets/AssetsTab.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import {useStudio} from "../../state/useStudio.ts";
+import { useStudio } from "../state/useStudio";
 import type { DragEvent } from "react";
 
 export function AssetsTab() {
