+++ packages/studio/src/layout/components/CanvasToolbar.tsx
