+++ packages/studio/src/layout/hooks/useFitScale.ts
