+++ packages/studio/src/state/useStudio.ts
@@ -1,84 +1,53 @@
-// src/state/useStudio.ts
-import { create } from "zustand";
-import { ProjectZ } from "../types/project.js";
-import type { Project } from "../types/project.js";
-import type { SchemaField } from "../types/schema.js";
+import { create } from 'zustand';
+import { ProjectZ } from '../types/project.js';
+import type { Project } from '../types/project.js';
 import {
   saveProjectToIDB,
   loadProjectFromIDB,
   deleteMissingAssetBlobs,
-} from "./storage.js";
-import jsonPatch from "fast-json-patch";
+} from '../assets/store/storage.js';
+import jsonPatch from 'fast-json-patch';
+import { createLayoutSlice, LayoutSlice, defaultCanvas } from '../layout/store';
+import { createDataSlice, DataSlice } from '../data/store';
+import { createAssetsSlice, AssetsSlice } from '../assets/store';
+import { createViewModelsSlice, ViewModelsSlice } from '../viewmodels/store';
+
 const { applyPatch, compare } = jsonPatch;
 
 export const defaultProject: Project = {
-  name: "Untitled",
-  version: "0.1",
-  layout: "<Grid/>",
+  name: 'Untitled',
+  version: '0.1',
+  layout: '<Grid/>',
   data: {},
   assets: [],
   screen: { width: 1280, height: 720 },
   meta: {
     assetFolders: [],
-    assetPaths: {}, // alias -> "Folder/Subfolder", пусто = корень
+    assetPaths: {},
   },
 };
 
-const defaultCanvas = { width: 1280, height: 720 };
+export type Tab = 'Layout' | 'Data' | 'ViewModels' | 'Assets';
+export type Dirty = { layout: boolean; data: boolean; assets: boolean };
 
-type Tab = "Layout" | "Data" | "ViewModels" | "Assets";
-type Dirty = { layout: boolean; data: boolean; assets: boolean };
+export type StudioState = LayoutSlice &
+  DataSlice &
+  AssetsSlice &
+  ViewModelsSlice & {
+    project: Project;
+    activeTab: Tab;
+    dirty: Dirty;
+    setTab: (t: Tab) => void;
+    hydrate: () => Promise<void>;
+    loadProject: (raw: unknown) => void;
+    exportProject: () => string;
+    newProject: () => void;
+    renameProject: (name: string) => void;
+    undo: () => void;
+    redo: () => void;
+  };
 
-type StudioState = {
-  project: Project;
-  activeTab: Tab;
-  dirty: Dirty;
-
-  selectAssetAliases: string[] | null;
-  clearSelectAssetAliases: () => void;
-
-  selectedSchema: string | null;
-  setSelectedSchema: (name: string | null) => void;
-  addSchema: () => void;
-  setSchemaFields: (name: string, fields: SchemaField[]) => void;
-
-  // tabs
-  setTab: (t: Tab) => void;
-
-  // persistence
-  hydrate: () => Promise<void>;         // загрузка из IndexedDB при старте
-  loadProject: (raw: unknown) => void;  // импорт из JSON
-  exportProject: () => string;          // экспорт в JSON
-  newProject: () => void;
-
-  // edits
-  renameProject: (name: string) => void;
-  setLayout: (s: string) => void;
-  setData: (o: any) => void;
-  setAssets: (a: Project["assets"]) => void;
-  addAssets: (a: Project["assets"]) => void;
-
-  // canvas
-  canvas: { width: number; height: number };
-  setCanvasSize: (w: number, h: number) => void;
-  swapCanvasSize: () => void;
-
-  // assets & folders
-  addAssetFolder: (path: string) => void;
-  setAssetPath: (alias: string, path: string | null) => void; // null => в корень
-  renameAssetDisplayName: (alias: string, name: string) => void; // только display name (НЕ alias)
-  deleteAssets: (aliases: string[]) => void;
-  deleteAsset: (alias: string) => void;
-
-  renameAssetFolder: (oldPath: string, newPath: string) => void; // переносит подпапки/ассеты
-  deleteAssetFolder: (path: string) => void; // удаляет папку(+вложенные), ассеты в корень
-
-  // undo/redo
-  undo: () => void;
-  redo: () => void;
-};
-
-export const useStudio = create<StudioState>((set, get) => {
+export const useStudio = create<StudioState>()((set, get) => {
   let saveTimer: ReturnType<typeof setTimeout> | undefined;
 
   const scheduleSave = () => {
@@ -89,7 +58,7 @@ export const useStudio = create<StudioState>((set, get) => {
         await saveProjectToIDB(p);
         await deleteMissingAssetBlobs((p.assets ?? []).map((a) => a.alias));
       } catch (e) {
-        console.warn("Save failed:", e);
+        console.warn('Save failed:', e);
       }
     }, 300);
   };
@@ -97,22 +66,17 @@ export const useStudio = create<StudioState>((set, get) => {
   type Command = { execute: () => void; undo: () => void };
   const history: Command[] = [];
   let cursor = 0;
-  const exec = (cmd: Command) => {
-    if (cursor < history.length) history.splice(cursor);
-    cmd.execute();
-    history.push(cmd);
-    if (history.length > 10) history.shift();
-    cursor = history.length;
-  };
+
   const runProjectCommand = (
     mutate: (p: Project) => Project,
-    sideEffects?: { onExecute?: () => void; onUndo?: () => void },
+    sideEffects?: { onExecute?: () => void; onUndo?: () => void }
   ) => {
     const before = structuredClone(get().project);
     const after = mutate(structuredClone(before));
     const patch = compare(before, after);
     const inverse = compare(after, before);
-    exec({
+    if (cursor < history.length) history.splice(cursor);
+    const cmd: Command = {
       execute() {
         set((s) => ({
           project: applyPatch(structuredClone(s.project), patch).newDocument,
@@ -129,36 +93,37 @@ export const useStudio = create<StudioState>((set, get) => {
         queueMicrotask(() => scheduleSave());
         sideEffects?.onUndo?.();
       },
-    });
+    };
+    cmd.execute();
+    history.push(cmd);
+    if (history.length > 10) history.shift();
+    cursor = history.length;
   };
 
   return {
-    project: { ...defaultProject }, // гидрируем позже
-    activeTab: "Layout",
+    project: { ...defaultProject },
+    activeTab: 'Layout',
     dirty: { layout: false, data: false, assets: false },
-    canvas: { ...defaultCanvas },
-    selectAssetAliases: null,
-    clearSelectAssetAliases: () => set({ selectAssetAliases: null }),
-    selectedSchema: null,
-    setSelectedSchema: (name) => set({ selectedSchema: name }),
+    ...createLayoutSlice(scheduleSave)(set, get),
+    ...createDataSlice(scheduleSave)(set, get),
+    ...createAssetsSlice(runProjectCommand)(set, get),
+    ...createViewModelsSlice()(set, get),
 
-    // ===== Tabs =====
     setTab: (t) => set({ activeTab: t }),
 
-    // ===== Persistence =====
     hydrate: async () => {
       try {
         const loaded = await loadProjectFromIDB();
         if (loaded) {
           set({
             project: loaded,
-            activeTab: "Layout",
+            activeTab: 'Layout',
             dirty: { layout: false, data: false, assets: false },
             canvas: { ...defaultCanvas },
           });
         }
       } catch (e) {
-        console.warn("Hydrate failed:", e);
+        console.warn('Hydrate failed:', e);
       }
     },
 
@@ -167,7 +132,7 @@ export const useStudio = create<StudioState>((set, get) => {
       set({
         project: parsed,
         dirty: { layout: false, data: false, assets: false },
-        activeTab: "Layout",
+        activeTab: 'Layout',
         canvas: { ...defaultCanvas },
       });
       scheduleSave();
@@ -184,204 +149,18 @@ export const useStudio = create<StudioState>((set, get) => {
       set({
         project: { ...defaultProject },
         dirty: { layout: false, data: false, assets: false },
-        activeTab: "Layout",
+        activeTab: 'Layout',
         canvas: { ...defaultCanvas },
       });
       scheduleSave();
     },
 
-    // ===== Edits =====
     renameProject: (name) => {
       const next = name.trim();
       if (!next) return;
       set((s) => ({ project: { ...s.project, name: next } }));
       scheduleSave();
     },
-
-    setLayout: (layout) => {
-      set((s) => ({
-        project: { ...s.project, layout },
-        dirty: { ...s.dirty, layout: true },
-      }));
-      scheduleSave();
-    },
-
-    setData: (data) => {
-      set((s) => ({
-        project: { ...s.project, data },
-        dirty: { ...s.dirty, data: true },
-      }));
-      scheduleSave();
-    },
-
-    addSchema: () => {
-      set((s) => {
-        const existing = new Set(Object.keys(s.project.data ?? {}));
-        const base = "New schema";
-        let name = base;
-        let i = 2;
-        while (existing.has(name)) name = `${base} ${i++}`;
-        const data = { ...s.project.data, [name]: [] as SchemaField[] };
-        return {
-          project: { ...s.project, data },
-          selectedSchema: name,
-          dirty: { ...s.dirty, data: true },
-        };
-      });
-      scheduleSave();
-    },
-
-    setSchemaFields: (name, fields) => {
-      set((s) => ({
-        project: {
-          ...s.project,
-          data: { ...s.project.data, [name]: fields },
-        },
-        dirty: { ...s.dirty, data: true },
-      }));
-      scheduleSave();
-    },
-
-    setAssets: (assets) =>
-      runProjectCommand((p) => ({ ...p, assets })),
-
-    addAssets: (toAdd) =>
-      runProjectCommand((p) => {
-        const assets = [...(p.assets ?? [])];
-        const byAlias = new Map(assets.map((a) => [a.alias, a]));
-        for (const asset of toAdd) {
-          const existing = byAlias.get(asset.alias);
-          if (existing) {
-            existing.src = asset.src;
-            if (!existing.name) existing.name = asset.name;
-          } else {
-            assets.push(asset);
-            byAlias.set(asset.alias, asset);
-          }
-        }
-        return { ...p, assets };
-      }),
-
-    // ===== Canvas =====
-    setCanvasSize: (width, height) =>
-      set((s) => {
-        const prev = s.project.screen ?? { width: 1280, height: 720 };
-        const w = Number.isFinite(width) ? Math.max(1, Math.round(width)) : prev.width;
-        const h = Number.isFinite(height) ? Math.max(1, Math.round(height)) : prev.height;
-        const next = { ...s.project, screen: { width: w, height: h } };
-        // сохраняем после обновления стейта
-        queueMicrotask(() => scheduleSave());
-        return { project: next };
-      }),
-
-    swapCanvasSize: () =>
-      set((s) => {
-        const prev = s.project.screen ?? { width: 1280, height: 720 };
-        const next = { ...s.project, screen: { width: prev.height, height: prev.width } };
-        queueMicrotask(() => scheduleSave());
-        return { project: next };
-      }),
-
-    // ===== Asset Folders & Paths =====
-    addAssetFolder: (path) =>
-      runProjectCommand((p) => {
-        const folders = new Set(p.meta?.assetFolders ?? []);
-        folders.add(path.trim());
-        return {
-          ...p,
-          meta: { ...(p.meta ?? {}), assetFolders: Array.from(folders) },
-        };
-      }),
-
-    setAssetPath: (alias, path) =>
-      runProjectCommand((p) => {
-        const meta = {
-          ...(p.meta ?? {}),
-          assetPaths: { ...(p.meta?.assetPaths ?? {}) },
-        };
-        if (!path || !path.trim()) {
-          delete meta.assetPaths[alias];
-        } else {
-          meta.assetPaths[alias] = path.trim();
-          const folders = new Set(meta.assetFolders ?? []);
-          folders.add(path.trim());
-          meta.assetFolders = Array.from(folders);
-        }
-        return { ...p, meta };
-      }),
-
-    renameAssetDisplayName: (alias, name) =>
-      runProjectCommand((p) => {
-        const assets = (p.assets ?? []).map((a) =>
-          a.alias === alias ? ({ ...a, name } as any) : a
-        );
-        return { ...p, assets };
-      }),
-
-    deleteAssets: (aliases) =>
-      runProjectCommand(
-        (p) => {
-          const remove = new Set(aliases);
-          const assets = (p.assets ?? []).filter((a) => !remove.has(a.alias));
-          const meta = {
-            ...(p.meta ?? {}),
-            assetPaths: { ...(p.meta?.assetPaths ?? {}) },
-          };
-          for (const alias of aliases) delete meta.assetPaths[alias];
-          return { ...p, assets, meta };
-        },
-        {
-          onExecute: () => set({ selectAssetAliases: [] }),
-          onUndo: () => set({ selectAssetAliases: aliases }),
-        },
-      ),
-
-    deleteAsset: (alias) => get().deleteAssets([alias]),
-
-    renameAssetFolder: (oldPath, newPath) =>
-      runProjectCommand((p) => {
-        const meta0 = p.meta ?? { assetFolders: [], assetPaths: {} };
-        const folders = (meta0.assetFolders ?? []).map((pth) =>
-          pth === oldPath || pth.startsWith(oldPath + "/")
-            ? newPath + pth.slice(oldPath.length)
-            : pth
-        );
-        const assetPaths = { ...(meta0.assetPaths ?? {}) };
-        for (const [alias, pth] of Object.entries(assetPaths)) {
-          if (!pth) continue;
-          if (pth === oldPath || pth.startsWith(oldPath + "/")) {
-            assetPaths[alias] = newPath + pth.slice(oldPath.length);
-          }
-        }
-        return {
-          ...p,
-          meta: {
-            ...meta0,
-            assetFolders: Array.from(new Set(folders)),
-            assetPaths,
-          },
-        };
-      }),
-
-    deleteAssetFolder: (path) =>
-      runProjectCommand((p) => {
-        const meta0 = p.meta ?? { assetFolders: [], assetPaths: {} };
-        const folders = (meta0.assetFolders ?? []).filter(
-          (pth) => pth !== path && !pth.startsWith(path + "/")
-        );
-        const assetPaths = { ...(meta0.assetPaths ?? {}) };
-        const assets = (p.assets ?? []).filter((a) => {
-          const pth = assetPaths[a.alias];
-          const inside = pth && (pth === path || pth.startsWith(path + "/"));
-          if (inside) delete assetPaths[a.alias];
-          return !inside;
-        });
-        return {
-          ...p,
-          assets,
-          meta: { ...meta0, assetFolders: folders, assetPaths },
-        };
-      }),
 
     undo: () => {
       if (cursor === 0) return;
