+++ packages/studio/src/viewmodels/store.ts
@@ -1,0 +1,14 @@
+import type { StateCreator } from 'zustand';
+
+export type ViewModelsSlice = {
+  // placeholder for future view model operations
+};
+
+export const createViewModelsSlice = (): StateCreator<
+  any,
+  [],
+  [],
+  ViewModelsSlice
+> => () => ({
+  // no view model state yet
+});
